requirements.txt
streamlit==1.40.0
streamlit-drawable-canvas
PyMuPDF
Pillow
streamlit
fitz